+++ src/completion/index.ts
@@ -1,2 +1,3 @@
 export { attributes } from './attributes';
 export { functions } from './functions';
+export { booleanSettings,booleanSettingNames, stringSettings, stringArraySettings } from './settings';