+++ src/just/index.ts
@@ -1,3 +1,3 @@
-export { execJust, getJustExecutable, setJustExecutable } from './exec';
+export { execJust, getJustExecutable, setJustExecutable, getJustShell, setJustShell, isJustUnstableFeaturesEnabled, enableJustUnstableFeatures, JustExecError } from './exec';
 export { getRecipes } from "./get-recipes";
 export { runRecipe, runRecipeOptions } from "./run-recipe";