+++ src/vscode/documentFormattingEditProvider.ts
@@ -1,7 +1,7 @@
 import { CancellationToken, DocumentFormattingEditProvider, FormattingOptions, OutputChannel, ProviderResult, Range, TextDocument, TextEdit } from "vscode";
 import { execJust } from "../just";
-import { open } from "fs";
 import { unlink, writeFile } from "fs/promises";
+import path = require("path");
 
 export class JustDocumentFormattingEditProvider implements DocumentFormattingEditProvider {
 
@@ -9,12 +9,25 @@ export class JustDocumentFormattingEditProvider implements DocumentFormattingEdi
     }
 
     provideDocumentFormattingEdits(document: TextDocument, options: FormattingOptions, token: CancellationToken): ProviderResult<TextEdit[]> {
-        return this.dumpJustfile(document);
+        const swapFileName = path.join(path.dirname(document.fileName), "._" + path.basename(document.fileName) + ".swap");
+        return this.dumpJustfile(document, swapFileName);
     }
 
-    async dumpJustfile(document: TextDocument): Promise<TextEdit[]> {
-        const swapFileName = document.fileName + ".swap";
-
+    async dumpJustfile(document: TextDocument, swapFileName: string): Promise<TextEdit[]> {
+        // try {
+        //     document.save();
+        //     const result = await execJust(['--justfile', document.fileName, '--fmt', '--check', '--unstable']);
+        //     if (result.exitCode === 0) {
+        //         // it is formatted correctly
+        //         return [];
+        //     }
+        // } catch (err) {
+        //     if (err instanceof JustExecError) {
+        //         if (err.kind == 'unformatted') {
+        //         }
+        //     }
+        //     this.outputChannel.appendLine("Failed to check file fmt: " + `, Error: ${err}`);
+        // }
         try {
             await writeFile(swapFileName, document.getText(), { flag: 'w+' });
         } catch (err) {
@@ -22,19 +35,21 @@ export class JustDocumentFormattingEditProvider implements DocumentFormattingEdi
             return [];
         }
 
-        const result = await execJust(['--justfile', swapFileName, '--dump', '--dump-format', 'just']);
+        try {
+            const result = await execJust(['--justfile', swapFileName, '--dump', '--dump-format', 'just']);
+            if (result.exitCode === 0) {
+                const wholeRange = new Range(
+                    document.positionAt(0),
+                    document.lineAt(document.lineCount - 1).range.end,
+                );
 
-        await unlink(swapFileName);
-
-        if (result.exitCode === 0) {
-            const wholeRange = new Range(
-                document.positionAt(0),
-                document.lineAt(document.lineCount - 1).range.end,
-            );
-
-            return [TextEdit.replace(wholeRange, result.stdout || '')];
+                return [TextEdit.replace(wholeRange, result.stdout || '')];
+            }
+        } catch (err) {
+            this.outputChannel.appendLine("Failed to format: " + document.fileName + `, Error: ${err}`);
         }
 
+        await unlink(swapFileName);
         return [];
     }
 }